% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{is_mle_constraint}
\alias{is_mle_constraint}
\title{Check if object is an mle_constraint}
\usage{
is_mle_constraint(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating whether \code{x} is an \code{mle_constraint}.
}
\description{
Check if object is an mle_constraint
}
\examples{
constraint <- mle_constraint(support = function(theta) all(theta > 0))
is_mle_constraint(constraint)  # TRUE
is_mle_constraint(list())      # FALSE
}
