% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{is_mle_problem}
\alias{is_mle_problem}
\title{Check if object is an mle_problem}
\usage{
is_mle_problem(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating whether \code{x} is an \code{mle_problem}.
}
\description{
Check if object is an mle_problem
}
\examples{
problem <- mle_problem(
  loglike = function(theta) -sum((theta - c(1, 2))^2)
)
is_mle_problem(problem)  # TRUE
is_mle_problem(list())   # FALSE
}
