% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mle_trace_data}
\alias{plot.mle_trace_data}
\title{Plot Trace Data Directly}
\usage{
\method{plot}{mle_trace_data}(x, ...)
}
\arguments{
\item{x}{An mle_trace_data object}

\item{...}{Arguments passed to plotting functions}
}
\value{
Called for side effects (generates a plot). Returns the input
  object invisibly.
}
\description{
Plot Trace Data Directly
}
