% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{print.mle_trace_data}
\alias{print.mle_trace_data}
\title{Print MLE Trace Data}
\usage{
\method{print}{mle_trace_data}(x, ...)
}
\arguments{
\item{x}{An mle_trace_data object.}

\item{...}{Additional arguments (unused).}
}
\value{
The input object, invisibly (for method chaining).
}
\description{
Print MLE Trace Data
}
